requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
openpyxl==3.1.2
pillow==10.1.0
pdf2image==1.16.3
opencv-python==4.8.1.78 